pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "td-rom"
version = "1.0.0"
description = "Python package for `Thermospheric Density Reduced Order Modeling`"
readme = "README.md"
authors = ["Vahid Nateghi, Matteo Manzi"]
license = "MIT"
repository = "https://github.com/PoincareTrajectories/td-rom"
homepage = "https://github.com/PoincareTrajectories/td-rom"

[tool.poetry.dependencies]
python = "^3.9"

typer = {extras = ["all"], version = "^0.7.0"}
rich = "^12.6.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.1"
black = {version = "^23.1.0", allow-prereleases = true}
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.12.0"}
mypy = "^0.991"
mypy-extensions = "^0.4.3"
pre-commit = "^3.1.1"
pydocstyle = "^6.1.1"
pylint = "^2.16.2"
pytest = "^6.2.5"
pyupgrade = "^3.2.2"
safety = "^2.1.1"
coverage = "^7.2.1"

coverage-badge = "^1.1.0"
pytest-html = "^3.2.0"
pytest-cov = "^4.0.0"
matplotlib = "^3.6.2"
pandas = "^1.5.3"
scikit-learn = "^1.1.3"
scipy = { version = "^1.9.3", python = ">=3.9,<3.12" }
ipykernel = "^6.21.2"
tensorflow-io-gcs-filesystem = { version = "^0.23.0", python = ">=3.9,<3.11" }
tensorflow-cpu = { version = "^2.9.3", python = ">=3.9,<3.11" }
pyJoules = "^0.5.1"
keras-tuner = "^1.1.3"

[tool.black]
# https://github.com/psf/black
target-version = ["py39"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 39
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.9
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
